chore: rename util file
